+++ sdk/src/client/bindings/python/iota_client/__init__.py
@@ -5,5 +5,7 @@ from .prefix_hex import *
 from .types.address import *
 from .types.common import *
 from .types.feature import *
+from .types.native_token import *
 from .types.output_id import *
+from .types.token_scheme import *
 from .types.unlock_condition import *
