+++ sdk/src/client/bindings/python/iota_client/types/native_token.py
@@ -1,0 +1,20 @@
+class NativeToken():
+    def __init__(self, id, amount):
+        """Initialise NativeToken
+
+        Parameters
+        ----------
+        id : string
+            Id of the token
+        amount : int
+            Native token amount
+        """
+        self.id = id
+        self.amount = amount
+
+    def as_dict(self):
+        config = dict(self.__dict__)
+
+        config['amount'] = str(hex(config['amount']))
+
+        return config
